moved static method "_interfaces_callback"
