+++ napalm_arubaoss/utils/__init__.py
@@ -3,7 +3,13 @@
 from napalm_arubaoss.utils.interfaces import (
     get_interface_details,
     get_interface_list,
-    fill_interface_dict
+    fill_interface_dict,
+    interfaces_callback
 )
 
-__all__ = ('get_interface_details', 'get_interface_list', 'fill_interface_dict')
+__all__ = (
+    'get_interface_details',
+    'get_interface_list',
+    'fill_interface_dict',
+    'interfaces_callback'
+)
